Schedule: avoid unnecesary refetchs and redraws.
Previously we issued a fetch and did a full redraw of the calendar
whenever we added, moved, or deleted an event and so even with a small number
of them everything started to behave like a crawling snail.

We still do that when somebody else updates the schedule from another session.