mediaedit: adds an id so relational works.
For new playlists if we do not set up an _id the relations point
to 'undefined' when saving. So we just put and _id and instruct
the server to generate a 'create' event. This is required because
otherwise when the browsers get an 'update' event they will not add
the new model to the corresponding collection.